requirements.txt
networkx
matplotlib
ipywidgets
qutip
cuda-quantum